Make it service driven
