+++ Interfaces/IInfoPage.cs
@@ -1,0 +1,6 @@
+﻿namespace BroadcastPluginSDK.Interfaces;
+
+public interface IInfoPage
+{
+    Control GetControl(); // Add this method to return a Control representation of the InfoPage.
+}
