+++ Interfaces/IProvider.cs
@@ -1,0 +1,6 @@
+﻿namespace BroadcastPluginSDK.Interfaces;
+
+public interface IProvider
+{
+    public event EventHandler<Dictionary<string, string>> DataReceived;
+}
